move global activation to extension-side
