+++ src/platform/serviceRegistry.web.ts
@@ -22,8 +22,7 @@ import { ProgressReporter } from './progress/progressReporter';
 import { StatusProvider } from './progress/statusProvider';
 import { IStatusProvider } from './progress/types';
 import { WorkspaceService } from './common/application/workspace.web';
-import { GlobalActivation } from './common/globalActivation';
-import { IExtensionSingleActivationService, IExtensionSyncActivationService } from './activation/types';
+import { IExtensionSyncActivationService } from './activation/types';
 import { OutputCommandListener } from './logging/outputCommandListener';
 import { ExportDialog } from './export/exportDialog';
 import { ExportFormat, IExport, IExportBase, IExportDialog, IFileConverter, INbConvertExport } from './export/types';
@@ -47,7 +46,6 @@ export function registerTypes(context: IExtensionContext, serviceManager: IServi
     serviceManager.addSingleton<IApplicationEnvironment>(IApplicationEnvironment, ApplicationEnvironment);
     serviceManager.addSingleton<IConfigurationService>(IConfigurationService, ConfigurationService);
     serviceManager.addSingleton<IStatusProvider>(IStatusProvider, StatusProvider);
-    serviceManager.addSingleton<IExtensionSingleActivationService>(IExtensionSingleActivationService, GlobalActivation);
     serviceManager.addSingleton<IDataScienceCommandListener>(IDataScienceCommandListener, OutputCommandListener);
     serviceManager.addSingleton<ExportFileOpener>(ExportFileOpener, ExportFileOpener);
     serviceManager.addSingleton<IExportBase>(IExportBase, ExportBase);
