+++ src/webviews/extension-side/globalActivation.ts
@@ -4,22 +4,22 @@
 import type { JSONObject } from '@lumino/coreutils';
 import { inject, injectable, multiInject, optional } from 'inversify';
 import * as vscode from 'vscode';
-import { ICommandManager, IDocumentManager, IWorkspaceService } from './application/types';
-import { PYTHON_FILE_ANY_SCHEME, PYTHON_LANGUAGE } from './constants';
-import { ContextKey } from './contextKey';
-import './extensions';
+import { ICommandManager, IDocumentManager, IWorkspaceService } from '../../platform/common/application/types';
+import { PYTHON_FILE_ANY_SCHEME, PYTHON_LANGUAGE } from '../../platform/common/constants';
+import { ContextKey } from '../../platform/common/contextKey';
+import '../../platform/common/extensions';
 import {
     IConfigurationService,
     IDataScienceCommandListener,
     IDisposable,
     IDisposableRegistry,
     IExtensionContext
-} from './types';
-import { debounceAsync, swallowExceptions } from './utils/decorators';
-import { noop } from './utils/misc';
+} from '../../platform/common/types';
+import { debounceAsync, swallowExceptions } from '../../platform/common/utils/decorators';
+import { noop } from '../../platform/common/utils/misc';
 import { sendTelemetryEvent } from '../../telemetry';
-import { EditorContexts, Telemetry } from './constants';
-import { IExtensionSingleActivationService } from '../activation/types';
+import { EditorContexts, Telemetry } from '../../platform/common/constants';
+import { IExtensionSingleActivationService } from '../../platform/activation/types';
 import { IDataScienceCodeLensProvider } from '../../interactive-window/editor-integration/types';
 import { IRawNotebookSupportedService } from '../../kernels/raw/types';
 import { hasCells } from '../../interactive-window/editor-integration/cellFactory';
